requirements.txt
Dash==1.12.0
pandas==1.0.3
numpy==1.18.1
requests==2.23.0
pyyaml==5.3.1
furl==2.1.0
dash-auth==1.3.2
python-dotenv==0.13.0
pytest==4.3.1
redis==3.5.3
python-redis-cache==1.1.0
